requirements.txt
pandas
apache-airflow-providers-snowflake==5.8.1
apache-airflow-providers-postgres==5.14.0